fixed compiling error

+++ src/game/editor/ed_editor.cpp
@@ -1862,7 +1862,7 @@ static void ExtractName(const char *pFileName, char *pName, int BufferSize)
 			pEnd = pFileName;
 	}
 
-	int Length = pEnd > pExtractedName ? min(BufferSize, pEnd-pExtractedName+1) : BufferSize;
+	int Length = pEnd > pExtractedName ? min(BufferSize, (int)(pEnd-pExtractedName+1)) : BufferSize;
 	str_copy(pName, pExtractedName, Length);
 }
